+++ server/impl/src/main/java/com/walmartlabs/concord/server/org/project/ProjectAccessManager.java
@@ -67,6 +67,12 @@ public class ProjectAccessManager {
             return e;
         }
 
+        if (level == ResourceAccessLevel.READER && (p.isGlobalReader() || p.isGlobalWriter())) {
+            return e;
+        } else if (level == ResourceAccessLevel.WRITER && p.isGlobalWriter()) {
+            return e;
+        }
+
         ProjectOwner owner = e.getOwner();
         if (owner != null && owner.getId().equals(p.getId())) {
             // the owner can do anything with his projects