+++ server/impl/src/main/java/com/walmartlabs/concord/server/org/secret/SecretManager.java
@@ -108,6 +108,12 @@ public class SecretManager {
             return e;
         }
 
+        if (level == ResourceAccessLevel.READER && (p.isGlobalReader() || p.isGlobalWriter())) {
+            return e;
+        } else if (level == ResourceAccessLevel.WRITER && p.isGlobalWriter()) {
+            return e;
+        }
+
         SecretOwner owner = e.getOwner();
         if (owner != null && owner.getId().equals(p.getId())) {
             // the owner can do anything with his secrets