+++ server/impl/src/main/java/com/walmartlabs/concord/server/security/apikey/ApiKeyRealm.java
@@ -72,7 +72,7 @@ public class ApiKeyRealm extends AuthorizingRealm {
                 .field("realm", REALM_NAME)
                 .log();
 
-        UserPrincipal p = new UserPrincipal(REALM_NAME, u.getId(), u.getName(), u.isAdmin(), u.getType());
+        UserPrincipal p = new UserPrincipal(REALM_NAME, u);
         return new SimpleAccount(Arrays.asList(p, t), t.getKey(), getName());
     }